+++ module/dramaticScene/torgeternityCombatTracker.js
@@ -47,6 +47,7 @@ export default class torgeternityCombatTracker extends foundry.applications.side
     context.conflictLine = combat?.conflictLineText;
     context.approvedActions = combat?.approvedActionsText;
     context.dsrLine = combat?.dsrText;
+    context.dramaRule = combat?.dramaRule;
 
     context.approved = {};
     if (combat)